Fix signed vs unsigned comparison warnings.

+++ src/undo.c
@@ -18,7 +18,7 @@
    Software Foundation, 59 Temple Place - Suite 330, Boston, MA
    02111-1307, USA.  */
 
-/*      $Id: undo.c,v 1.19 2005/01/27 01:33:19 rrt Exp $        */
+/*      $Id: undo.c,v 1.20 2005/01/29 12:21:32 rrt Exp $        */
 
 #include "config.h"
 
@@ -92,7 +92,7 @@ void undo_save(int type, Point pt, size_t arg1, size_t arg2)
  */
 static Undo *revert_action(Undo *up)
 {
-  int i;
+  size_t i;
 
   doing_undo = TRUE;
